+++ ext4/tar2ext4/tar2ext4.go
@@ -77,7 +77,7 @@ func MaximumDiskSize(size int64) Option {
 
 // FreeSpace instructs the writer to add additional free space to the disk.
 // If not provided, then 0GB is the default.
-func FreeSpace(size uint32) Option {
+func FreeSpace(size int64) Option {
 	return func(p *params) {
 		p.ext4opts = append(p.ext4opts, compactext4.FreeSpace(size))
 	}
